Fixing critical genotype determination bug
Heterozygous positions were not being scored due to the mclapply update. This bug has been in place since the August 12th update. The bug caused excessive genotype ambiguity due to the ignorance of heterozygous SNPs. If you are using a post August12th version of PING it is essential to update to fix this bug!